Add files via upload
A decentralized lottery system is implemented using solidity programming language in which users can participate in lottery.
It provide equal chance to all the participants to win as it is not controlled by a single entity . Smart contract uses algorithm that picks up the winner and deposit money into participant's wallet.The code is visible to all the participants involved in a lottery. After, the winner is picked up , our smart contract removes all the participants from the lottery.